Reverted workaround. Whatever the problem was, it wasn't in this source code.
[git-p4: depot-paths = "//depot/openwatcom/": change = 27244]

+++ bld/pgchart/c/pgscale.c
@@ -24,8 +24,7 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  Chart scaling routines.
 *
 ****************************************************************************/
 
@@ -134,7 +133,7 @@ static float scale_ticinterval( axistype _WCI86FAR *axis, float max, float min )
     axis->ticdecimals = tic_decimals( interval[ s ] );
 
     /* set step increment (adjust to precision) */
-    ds = (float)1.0;
+    ds = 1.0;
     for( i = 0; i < axis->ticdecimals; ++i ) {
         ds = ds * 10;
     }
